Roles auth

+++ Group5/routes/web.php
@@ -22,16 +22,16 @@ Route::get('/', function () {
     return view('index');
 });
 
-Route::middleware([
-    'auth:sanctum',
-    config('jetstream.auth_session'),
-    'verified',
-])->group(function () {
-    Route::get('/dashboard', function () {
-        $users = User::all();
-        return view('dashboard', compact('users'));
-    })->name('dashboard');
-});
+// Route::middleware([
+//     'auth:sanctum',
+//     config('jetstream.auth_session'),
+//     'verified',
+// ])->group(function () {
+//     Route::get('/dashboard', function () {
+//         $users = User::all();
+//         return view('dashboard', compact('users'));
+//     })->name('dashboard');
+// });
 Route::get('/contact', function () {
     $users = User::all();
     return view('contact');
@@ -52,18 +52,23 @@ Route::get('/register-view', function(){
     return view('auth.register');
 });
 
-//users CRUD
-Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update');
-Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show');
-Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy');
+//Admin CRUD
+Route::get('/dashboard', function () {
+    $users = User::all();
+    return view('dashboard', compact('users'));
+})->name('dashboard')->middleware('role:admin');
+Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit')->middleware('role:admin');
+Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update')->middleware('role:admin');
+Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show')->middleware('role:admin');
+Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy')->middleware('role:admin');
 
 
 //Employer
-Route::get('/jobs', [JobController::class, 'index'])->name('job.index');
-Route::get('/jobs/create', [JobController::class, 'create'])->name('job.create');
-Route::post('/store', [JobController::class, 'store'])->name('job.store');
+Route::get('/jobs', [JobController::class, 'index'])->name('job.index')->middleware('role:employer');
+Route::get('/jobs/create', [JobController::class, 'create'])->name('job.create')->middleware('role:employer');
+Route::post('/store', [JobController::class, 'store'])->name('job.store')->middleware('role:employer');
 Route::get('/job/{job}/edit', [JobController::class, 'edit'])->name('job.edit');
+
 Route::put('/job/{job}/update', [JobController::class, 'update'])->name('job.update');
 Route::delete('/job/{job}/delete', [JobController::class, 'delete'])->name('job.delete');
 Route::get('/job/{job}/applicants', [JobController::class, 'show_applicants'])->name('job.applicants');
@@ -75,8 +80,9 @@ Route::post('/applicant/submit-application', [ApplicationController::class, 'sub
 Route::get('/applicant/my-applications', [ApplicationController::class, 'my_applications'])->name('applicant.my-applications');
 Route::delete('/applicant/{application}/cancel', [ApplicationController::class, 'cancel'])->name('applicant.cancel');
 
-Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat');
-Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store');
-Route::get('/category/edit/{id}', [CategoryController::class, 'Edit']);
-Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category');
-Route::get('/category/delete/{id}', [CategoryController::class, 'Delete']);
+
+Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat')->middleware('role:admin');
+Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store')->middleware('role:admin');
+Route::get('/category/edit/{id}', [CategoryController::class, 'Edit'])->middleware('role:admin');
+Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category')->middleware('role:admin');
+Route::get('/category/delete/{id}', [CategoryController::class, 'Delete'])->middleware('role:admin');
